perplexity/requirements.txt
flask
google-generativeai
PyPDF2
